Updated style, html structure, and js functionality
The style was updated to fit most screens. It is not perfect but it looks
better. The html added a new div. The javascript needed to be fixed with
the width and height of the grid to conform to the 760 px size desired
so a const of 2 was used to mimic the border shared vertically and
horizontally in the tiles so it takes those pixels off of the tile.